requirements.txt
gradio
gitpython
openai
python-dotenv
beautifulsoup4
